Record region stats
